refactor(cli): extract runtime loop orchestration module (#178)
