+++ crates/pi-coding-agent/src/main.rs
@@ -6,6 +6,7 @@ mod diagnostics_commands;
 mod events;
 mod github_issues;
 mod macro_profile_commands;
+mod runtime_loop;
 mod session;
 mod session_commands;
 mod session_navigation_commands;
@@ -19,13 +20,9 @@ mod trust_roots;
 
 use std::{
     collections::{BTreeMap, HashMap, HashSet},
-    future::Future,
-    io::{Read, Write},
+    io::Write,
     path::{Path, PathBuf},
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc, Mutex,
-    },
+    sync::{Arc, Mutex},
     time::{Duration, Instant, SystemTime, UNIX_EPOCH},
 };
 
@@ -40,7 +37,6 @@ use pi_ai::{
 };
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use tokio::io::{AsyncBufReadExt, BufReader};
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::EnvFilter;
 
@@ -96,6 +92,10 @@ pub(crate) use crate::macro_profile_commands::{
     render_profile_show, save_macro_file, save_profile_store, validate_macro_command_entry,
     validate_macro_name, validate_profile_name, MacroCommand, MacroFile, ProfileCommand,
     ProfileStoreFile, MACRO_SCHEMA_VERSION, MACRO_USAGE, PROFILE_SCHEMA_VERSION, PROFILE_USAGE,
+};
+pub(crate) use crate::runtime_loop::{
+    resolve_prompt_input, run_interactive, run_prompt, run_prompt_with_cancellation,
+    InteractiveRuntimeConfig, PromptRunStatus,
 };
 use crate::session::{SessionImportMode, SessionStore};
 #[cfg(test)]
@@ -1399,13 +1399,6 @@ struct ProfileDefaults {
     auth: ProfileAuthDefaults,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum PromptRunStatus {
-    Completed,
-    Cancelled,
-    TimedOut,
-}
-
 #[derive(Debug, Clone, Copy)]
 struct RenderOptions {
     stream_output: bool,
@@ -1457,13 +1450,6 @@ fn build_auth_command_config(cli: &Cli) -> AuthCommandConfig {
         anthropic_auth_mode: cli.anthropic_auth_mode.into(),
         google_auth_mode: cli.google_auth_mode.into(),
     }
-}
-
-#[derive(Clone, Copy)]
-struct InteractiveRuntimeConfig<'a> {
-    turn_timeout_ms: u64,
-    render_options: RenderOptions,
-    command_context: CommandExecutionContext<'a>,
 }
 
 #[derive(Clone)]
@@ -2107,35 +2093,6 @@ async fn main() -> Result<()> {
     run_interactive(agent, session_runtime, interactive_config).await
 }
 
-fn resolve_prompt_input(cli: &Cli) -> Result<Option<String>> {
-    if let Some(prompt) = &cli.prompt {
-        return Ok(Some(prompt.clone()));
-    }
-
-    let Some(path) = cli.prompt_file.as_ref() else {
-        return Ok(None);
-    };
-
-    if path == std::path::Path::new("-") {
-        let mut prompt = String::new();
-        std::io::stdin()
-            .read_to_string(&mut prompt)
-            .context("failed to read prompt from stdin")?;
-        return Ok(Some(ensure_non_empty_text(
-            prompt,
-            "stdin prompt".to_string(),
-        )?));
-    }
-
-    let prompt = std::fs::read_to_string(path)
-        .with_context(|| format!("failed to read prompt file {}", path.display()))?;
-
-    Ok(Some(ensure_non_empty_text(
-        prompt,
-        format!("prompt file {}", path.display()),
-    )?))
-}
-
 fn validate_github_issues_bridge_cli(cli: &Cli) -> Result<()> {
     if !cli.github_issues_bridge {
         return Ok(());
@@ -2493,164 +2450,6 @@ fn initialize_session(agent: &mut Agent, cli: &Cli, system_prompt: &str) -> Resu
     }
 
     Ok(SessionRuntime { store, active_head })
-}
-
-async fn run_prompt(
-    agent: &mut Agent,
-    session_runtime: &mut Option<SessionRuntime>,
-    prompt: &str,
-    turn_timeout_ms: u64,
-    render_options: RenderOptions,
-) -> Result<()> {
-    let status = run_prompt_with_cancellation(
-        agent,
-        session_runtime,
-        prompt,
-        turn_timeout_ms,
-        tokio::signal::ctrl_c(),
-        render_options,
-    )
-    .await?;
-    if status == PromptRunStatus::Cancelled {
-        println!("\nrequest cancelled\n");
-    } else if status == PromptRunStatus::TimedOut {
-        println!("\nrequest timed out\n");
-    }
-    Ok(())
-}
-
-async fn run_interactive(
-    mut agent: Agent,
-    mut session_runtime: Option<SessionRuntime>,
-    config: InteractiveRuntimeConfig<'_>,
-) -> Result<()> {
-    let stdin = BufReader::new(tokio::io::stdin());
-    let mut lines = stdin.lines();
-
-    loop {
-        print!("pi> ");
-        std::io::stdout()
-            .flush()
-            .context("failed to flush stdout")?;
-
-        let Some(line) = lines.next_line().await? else {
-            break;
-        };
-
-        let trimmed = line.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-
-        if trimmed.starts_with('/') {
-            if handle_command_with_session_import_mode(
-                trimmed,
-                &mut agent,
-                &mut session_runtime,
-                config.command_context.tool_policy_json,
-                config.command_context.session_import_mode,
-                config.command_context.profile_defaults,
-                config.command_context.skills_command_config,
-                config.command_context.auth_command_config,
-            )? == CommandAction::Exit
-            {
-                break;
-            }
-            continue;
-        }
-
-        let status = run_prompt_with_cancellation(
-            &mut agent,
-            &mut session_runtime,
-            trimmed,
-            config.turn_timeout_ms,
-            tokio::signal::ctrl_c(),
-            config.render_options,
-        )
-        .await?;
-        if status == PromptRunStatus::Cancelled {
-            println!("\nrequest cancelled\n");
-        } else if status == PromptRunStatus::TimedOut {
-            println!("\nrequest timed out\n");
-        }
-    }
-
-    Ok(())
-}
-
-async fn run_prompt_with_cancellation<F>(
-    agent: &mut Agent,
-    session_runtime: &mut Option<SessionRuntime>,
-    prompt: &str,
-    turn_timeout_ms: u64,
-    cancellation_signal: F,
-    render_options: RenderOptions,
-) -> Result<PromptRunStatus>
-where
-    F: Future,
-{
-    let checkpoint = agent.messages().to_vec();
-    let streamed_output = Arc::new(AtomicBool::new(false));
-    let stream_delta_handler = if render_options.stream_output {
-        let streamed_output = streamed_output.clone();
-        let stream_delay_ms = render_options.stream_delay_ms;
-        Some(Arc::new(move |delta: String| {
-            if delta.is_empty() {
-                return;
-            }
-            streamed_output.store(true, Ordering::Relaxed);
-            print!("{delta}");
-            let _ = std::io::stdout().flush();
-            if stream_delay_ms > 0 {
-                std::thread::sleep(Duration::from_millis(stream_delay_ms));
-            }
-        }) as StreamDeltaHandler)
-    } else {
-        None
-    };
-    tokio::pin!(cancellation_signal);
-
-    enum PromptOutcome<T> {
-        Result(T),
-        Cancelled,
-        TimedOut,
-    }
-
-    let prompt_result = if turn_timeout_ms == 0 {
-        tokio::select! {
-            result = agent.prompt_with_stream(prompt, stream_delta_handler.clone()) => PromptOutcome::Result(result),
-            _ = &mut cancellation_signal => PromptOutcome::Cancelled,
-        }
-    } else {
-        let timeout = tokio::time::sleep(Duration::from_millis(turn_timeout_ms));
-        tokio::pin!(timeout);
-        tokio::select! {
-            result = agent.prompt_with_stream(prompt, stream_delta_handler.clone()) => PromptOutcome::Result(result),
-            _ = &mut cancellation_signal => PromptOutcome::Cancelled,
-            _ = &mut timeout => PromptOutcome::TimedOut,
-        }
-    };
-
-    let prompt_result = match prompt_result {
-        PromptOutcome::Result(result) => result,
-        PromptOutcome::Cancelled => {
-            agent.replace_messages(checkpoint);
-            return Ok(PromptRunStatus::Cancelled);
-        }
-        PromptOutcome::TimedOut => {
-            agent.replace_messages(checkpoint);
-            return Ok(PromptRunStatus::TimedOut);
-        }
-    };
-
-    let new_messages = prompt_result?;
-    persist_messages(session_runtime, &new_messages)?;
-    print_assistant_messages(
-        &new_messages,
-        render_options,
-        streamed_output.load(Ordering::Relaxed),
-    );
-    Ok(PromptRunStatus::Completed)
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
